+++ apps/server/src/modules/common-cartridge/import/utils/common-cartridge-organization-visitor.spec.ts
@@ -1,13 +1,10 @@
-/*
+import { DEFAULT_FILE_PARSER_OPTIONS } from '@modules/common-cartridge';
 import AdmZip from 'adm-zip';
-import { readFile } from 'fs/promises';
 import { load } from 'cheerio';
-import { CommonCartridgeOrganizationProps, DEFAULT_FILE_PARSER_OPTIONS } from '../common-cartridge-import.types';
+import { readFile } from 'fs/promises';
 import { CommonCartridgeOrganizationVisitor } from './common-cartridge-organization-visitor';
 
-
-TODO refactor for new visitor
-describe('CommonCartridgeOrganizationVisitor', () => {
+describe(CommonCartridgeOrganizationVisitor.name, () => {
 	const setupDocument = async (loadFile: boolean) => {
 		if (!loadFile) {
 			const document = load('<manifest></manifest>', { xml: true });
@@ -24,15 +21,11 @@ describe('CommonCartridgeOrganizationVisitor', () => {
 		return document;
 	};
 
-	describe('findAllOrganizations', () => {
+	describe('findAllNodes', () => {
 		describe('when organizations are present', () => {
 			const setup = async () => {
 				const document = await setupDocument(true);
-				const sut = new CommonCartridgeOrganizationVisitor(document, {
-					maxSearchDepth: 1,
-					pathSeparator: DEFAULT_FILE_PARSER_OPTIONS.pathSeparator,
-					inputFormat: DEFAULT_FILE_PARSER_OPTIONS.inputFormat,
-				});
+				const sut = new CommonCartridgeOrganizationVisitor(document, DEFAULT_FILE_PARSER_OPTIONS);
 
 				return { sut };
 			};
@@ -40,22 +33,9 @@ describe('CommonCartridgeOrganizationVisitor', () => {
 			it('should return the organizations', async () => {
 				const { sut } = await setup();
 
-				const result = sut.findAllOrganizations();
+				const result = sut.findAllNodes();
 
 				expect(result).toHaveLength(117);
-				result.forEach((organization) => {
-					expect(organization).toEqual<CommonCartridgeOrganizationProps>({
-						identifier: expect.any(String),
-						identifierRef: expect.any(String),
-						title: expect.any(String),
-						path: expect.any(String),
-						pathDepth: expect.any(Number),
-						isResource: expect.any(Boolean),
-						isInlined: expect.any(Boolean),
-						resourcePath: expect.any(String),
-						resourceType: expect.any(String),
-					});
-				});
 			});
 		});
 
@@ -70,11 +50,10 @@ describe('CommonCartridgeOrganizationVisitor', () => {
 			it('should return an empty array', async () => {
 				const { sut } = await setup();
 
-				const result = sut.findAllOrganizations();
+				const result = sut.findAllNodes();
 
 				expect(result).toHaveLength(0);
 			});
 		});
 	});
 });
-*/
